sys/hashes, sys/checksum: mv fletcher* checksum
