+++ sys/checksum/fletcher16.c
@@ -1,0 +1,40 @@
+/*
+ * Copyright 2015 Eistec AB
+ *
+ * This file is subject to the terms and conditions of the GNU Lesser
+ * General Public License v2.1. See the file LICENSE in the top level
+ * directory for more details.
+ */
+
+/**
+ * @ingroup     sys_checksum_fletcher16
+ * @{
+ *
+ * @file
+ * @brief       Fletcher16 implementation
+ *
+ * @author      Joakim Nohlgård <joakim.nohlgard@eistec.se>
+ *
+ * @}
+ */
+
+#include "checksum/fletcher16.h"
+
+uint16_t fletcher16(const uint8_t *data, size_t bytes)
+{
+    uint16_t sum1 = 0xff, sum2 = 0xff;
+
+    while (bytes) {
+        size_t tlen = bytes > 20 ? 20 : bytes;
+        bytes -= tlen;
+        do {
+            sum2 += sum1 += *data++;
+        } while (--tlen);
+        sum1 = (sum1 & 0xff) + (sum1 >> 8);
+        sum2 = (sum2 & 0xff) + (sum2 >> 8);
+    }
+    /* Second reduction step to reduce sums to 8 bits */
+    sum1 = (sum1 & 0xff) + (sum1 >> 8);
+    sum2 = (sum2 & 0xff) + (sum2 >> 8);
+    return (sum2 << 8) | sum1;
+}
