+++ sys/hashes/hashes.c
@@ -12,7 +12,6 @@
  * @file
  * @author      Jason Linehan <patientulysses@gmail.com>
  * @author      Christian Mehlis <mehlis@inf.fu-berlin.de>
- * @author      Joakim Nohlgård <joakim.nohlgard@eistec.se>
  */
 
 #include "hashes.h"
@@ -109,42 +108,4 @@ uint32_t one_at_a_time_hash(const uint8_t *buf, size_t len)
     hash ^= hash >> 11;
     hash += hash << 15;
     return hash;
-}
-
-uint16_t fletcher16(const uint8_t *data, size_t bytes)
-{
-    uint16_t sum1 = 0xff, sum2 = 0xff;
-
-    while (bytes) {
-        size_t tlen = bytes > 20 ? 20 : bytes;
-        bytes -= tlen;
-        do {
-            sum2 += sum1 += *data++;
-        } while (--tlen);
-        sum1 = (sum1 & 0xff) + (sum1 >> 8);
-        sum2 = (sum2 & 0xff) + (sum2 >> 8);
-    }
-    /* Second reduction step to reduce sums to 8 bits */
-    sum1 = (sum1 & 0xff) + (sum1 >> 8);
-    sum2 = (sum2 & 0xff) + (sum2 >> 8);
-    return (sum2 << 8) | sum1;
-}
-
-uint32_t fletcher32(const uint16_t *data, size_t words)
-{
-    uint32_t sum1 = 0xffff, sum2 = 0xffff;
-
-    while (words) {
-        unsigned tlen = words > 359 ? 359 : words;
-        words -= tlen;
-        do {
-            sum2 += sum1 += *data++;
-        } while (--tlen);
-        sum1 = (sum1 & 0xffff) + (sum1 >> 16);
-        sum2 = (sum2 & 0xffff) + (sum2 >> 16);
-    }
-    /* Second reduction step to reduce sums to 16 bits */
-    sum1 = (sum1 & 0xffff) + (sum1 >> 16);
-    sum2 = (sum2 & 0xffff) + (sum2 >> 16);
-    return (sum2 << 16) | sum1;
 }
